Rename files so they match the ones they test
